requirements.txt
yfinance>=0.2.36
pykrx>=1.0.45
Jinja2>=3.1.3
PyYAML>=6.0.1
matplotlib>=3.8.0
pandas>=2.0.0
numpy>=1.24.0
